move varint32 sttuff to cpp
